Update i3status
Changed the time interval from 5 seconds to 30 seconds, so as to prevent too many context switches on the CPU, whilst still maintaining up to date information.
Alongside this, I changed the path for the cpu_temperature field as I found that for some reason the directory can change between restarts. Fortunately, it seems that the only thing that changes is a single number in a folder directory. The hardcoded '4' was replaced with a wildcard '*'. Nice.